Move Authenticator into SecurityModule
