+++ integration/src/test/scala/org/knora/webapi/core/LayersTest.scala
@@ -45,7 +45,6 @@ import org.knora.webapi.slice.common.api.*
 import org.knora.webapi.slice.common.repo.service.PredicateObjectMapper
 import org.knora.webapi.slice.infrastructure.CacheManager
 import org.knora.webapi.slice.infrastructure.InfrastructureModule
-import org.knora.webapi.slice.infrastructure.ScopeResolver
 import org.knora.webapi.slice.infrastructure.api.ManagementEndpoints
 import org.knora.webapi.slice.infrastructure.api.ManagementRoutes
 import org.knora.webapi.slice.ontology.api.service.RestCardinalityService
@@ -65,6 +64,8 @@ import org.knora.webapi.slice.resources.repo.service.ResourcesRepo
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepoLive
 import org.knora.webapi.slice.search.api.SearchApiRoutes
 import org.knora.webapi.slice.search.api.SearchEndpoints
+import org.knora.webapi.slice.security.ScopeResolver
+import org.knora.webapi.slice.security.SecurityModule
 import org.knora.webapi.store.iiif.IIIFRequestMessageHandler
 import org.knora.webapi.store.iiif.IIIFRequestMessageHandlerLive
 import org.knora.webapi.store.iiif.api.SipiService
@@ -98,7 +99,6 @@ object LayersTest {
     ApiRoutes &
     AppRouter &
     AssetPermissionsResponder &
-    Authenticator &
     AuthorizationRestService &
     CardinalityHandler &
     CardinalityService &
@@ -129,7 +129,7 @@ object LayersTest {
     ResourceUtilV2 &
     ResourcesRepo &
     RestCardinalityService &
-    ScopeResolver &
+    SecurityModule.Provided &
     SearchApiRoutes &
     SearchEndpoints &
     SearchResponderV2Module.Provided &
@@ -160,7 +160,6 @@ object LayersTest {
       ApiRoutes.layer,
       AppRouter.layer,
       AssetPermissionsResponder.layer,
-      AuthenticatorLive.layer,
       AuthorizationRestService.layer,
       BaseEndpoints.layer,
       CardinalityHandler.layer,
@@ -193,7 +192,7 @@ object LayersTest {
       ResourceUtilV2Live.layer,
       ResourcesRepoLive.layer,
       RestCardinalityServiceLive.layer,
-      ScopeResolver.layer,
+      SecurityModule.layer,
       SearchApiRoutes.layer,
       SearchEndpoints.layer,
       SearchResponderV2Module.layer,
