+++ integration/src/test/scala/org/knora/webapi/e2e/v2/AuthenticationV2E2ESpec.scala
@@ -16,9 +16,9 @@ import org.knora.webapi.E2ESpec
 import org.knora.webapi.messages.store.triplestoremessages.TriplestoreJsonProtocol
 import org.knora.webapi.messages.v2.routing.authenticationmessages.AuthenticationV2JsonProtocol
 import org.knora.webapi.messages.v2.routing.authenticationmessages.LoginResponse
-import org.knora.webapi.routing.Authenticator
 import org.knora.webapi.routing.UnsafeZioRun
 import org.knora.webapi.sharedtestdata.SharedTestDataADM
+import org.knora.webapi.slice.security.Authenticator
 import org.knora.webapi.util.MutableTestString
 
 import pekko.http.scaladsl.model.*
