+++ webapi/src/main/scala/org/knora/webapi/routing/KnoraRoute.scala
@@ -15,6 +15,7 @@ import scala.concurrent.Future
 
 import org.knora.webapi.config.AppConfig
 import org.knora.webapi.messages.StringFormatter
+import org.knora.webapi.slice.security.Authenticator
 
 import pekko.actor.ActorRef
 import pekko.actor.ActorSystem
