+++ webapi/src/main/scala/org/knora/webapi/slice/common/api/BaseEndpoints.scala
@@ -27,10 +27,10 @@ import dsp.errors.*
 import org.knora.webapi.messages.util.KnoraSystemInstances.Users.AnonymousUser
 import org.knora.webapi.messages.v2.routing.authenticationmessages.CredentialsIdentifier
 import org.knora.webapi.messages.v2.routing.authenticationmessages.KnoraCredentialsV2.KnoraPasswordCredentialsV2
-import org.knora.webapi.routing.Authenticator
 import org.knora.webapi.routing.UnsafeZioRun
 import org.knora.webapi.slice.admin.domain.model.Email
 import org.knora.webapi.slice.admin.domain.model.User
+import org.knora.webapi.slice.security.Authenticator
 
 final case class BaseEndpoints(authenticator: Authenticator)(implicit val r: zio.Runtime[Any]) {
